PDFBOX-2576: use simple data types instead of COSInteger to handle object and generation number of a COSObject
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1663131 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/cos/COSObjectKey.java
@@ -34,7 +34,7 @@ public class COSObjectKey implements Comparable<COSObjectKey>
      */
     public COSObjectKey(COSObject object)
     {
-        this( object.getObjectNumber().intValue(), object.getGenerationNumber().intValue() );
+        this(object.getObjectNumber(), object.getGenerationNumber());
     }
 
     /**
